backend/requirements-simple.txt
# Core Framework
fastapi
uvicorn
pydantic
pydantic-settings

# Database
sqlalchemy
alembic
psycopg2-binary

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# External APIs
openai
requests

# Background Tasks
celery
redis

# File Processing
pandas
openpyxl

# Testing
pytest
pytest-asyncio